chore: migrated tests for solutions to the test directory.
