+++ .idea/libraries/Dart_SDK.xml
@@ -1,27 +1,25 @@
 <component name="libraryTable">
   <library name="Dart SDK">
     <CLASSES>
-      <root url="file:///usr/local/opt/dart/libexec/lib/async" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/cli" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/collection" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/convert" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/core" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/developer" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/ffi" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/html" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/indexed_db" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/io" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/isolate" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/js" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/js_util" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/math" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/mirrors" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/svg" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/typed_data" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/wasm" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/web_audio" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/web_gl" />
-      <root url="file:///usr/local/opt/dart/libexec/lib/web_sql" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/async" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/cli" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/collection" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/convert" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/core" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/developer" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/ffi" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/html" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/indexed_db" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/io" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/isolate" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/js" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/js_util" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/math" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/mirrors" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/svg" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/typed_data" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/web_audio" />
+      <root url="file://$USER_HOME$/Development/flutter/bin/cache/dart-sdk/lib/web_gl" />
     </CLASSES>
     <JAVADOC />
     <SOURCES />
