+++ lib/leetcode/arrays/max_consecutive_ones.dart
@@ -1,6 +1,4 @@
-import 'package:test/test.dart';
-
-/* 
+/*
 Given a binary array nums, return the maximum number of consecutive 1's in the array.
 Constraints:
 1 <= nums.length <= 105
@@ -30,16 +28,4 @@ int maxConsecutiveOnes(List<int> array) {
 
     return maxCount;
   }
-}
-
-void main() {
-  test('maxConsecutiveOnes([1, 1, 0, 1, 1, 1]) should return 3', () {
-    final result = maxConsecutiveOnes([1, 1, 0, 1, 1, 1]);
-    expect(result, 3);
-  });
-
-  test('maxConsecutiveOnes([1, 0, 1, 1, 0, 1]) should return 2', () {
-    final result = maxConsecutiveOnes([1, 0, 1, 1, 0, 1]);
-    expect(result, 2);
-  });
 }
