+++ lib/project_euler/problem_1.dart
@@ -1,5 +1,3 @@
-import 'package:test/test.dart';
-
 /*
 Problem 1:
 [https://projecteuler.net/problem=1] or [https://www.freecodecamp.org/learn/coding-interview-prep/project-euler/problem-1-multiples-of-3-and-5]
@@ -24,26 +22,3 @@ int multiplesOf3and5(int number) {
   return sumOfMultiples;
 }
 
-// Tests
-void main() {
-  test('multiplesOf3and5(10) should return a number.', () {
-    final result = multiplesOf3and5(10);
-    expect(result, 23);
-  });
-  test('multiplesOf3and5(49) should return 543.', () {
-    final result = multiplesOf3and5(49);
-    expect(result, 543);
-  });
-  test('multiplesOf3and5(1000) should return 233168.', () {
-    final result = multiplesOf3and5(1000);
-    expect(result, 233168);
-  });
-  test('multiplesOf3and5(8456) should return 16687353.', () {
-    final result = multiplesOf3and5(8456);
-    expect(result, 16687353);
-  });
-  test('multiplesOf3and5(19564) should return 89301183.', () {
-    final result = multiplesOf3and5(19564);
-    expect(result, 89301183);
-  });
-}
