+++ lib/project_euler/problem_2.dart
@@ -1,5 +1,3 @@
-import 'package:test/test.dart';
-
 /*
 Problem 2:
 [https://projecteuler.net/problem=2] or [https://www.freecodecamp.org/learn/coding-interview-prep/project-euler/problem-2-even-fibonacci-numbers]
@@ -64,37 +62,4 @@ int evenFibonacciSumRecursion(int n) {
   } else {
     return evenFibonacciSumRecursion(n - 1) + evenFibonacciSumRecursion(n - 2);
   }
-}
-
-// Tests
-void main() {
-  // using the loop solution
-  test('evenFibonacciSum(8) should return 10', () {
-    final result = evenFibonacciSum(8);
-    expect(result, 10);
-  });
-  test('evenFibonacciSum(10) should return 10', () {
-    final result = evenFibonacciSum(10);
-    expect(result, 10);
-  });
-  test('evenFibonacciSum(34) should return 44', () {
-    final result = evenFibonacciSum(34);
-    expect(result, 44);
-  });
-  test('evenFibonacciSum(60) should return 44', () {
-    final result = evenFibonacciSum(60);
-    expect(result, 44);
-  });
-  test('evenFibonacciSum(1000) should return 798', () {
-    final result = evenFibonacciSum(1000);
-    expect(result, 798);
-  });
-  test('evenFibonacciSum(100000) should return 60696', () {
-    final result = evenFibonacciSum(100000);
-    expect(result, 60696);
-  });
-  test('evenFibonacciSum(4000 000) should return 4613732', () {
-    final result = evenFibonacciSum(4000000);
-    expect(result, 4613732);
-  });
 }
